config/config.py
from dataclasses import dataclass
from google.adk.tools import google_search

@dataclass
class NewsEngineConfig:
    agent_name: str = "NewsFindingAgent"
    description: str = "Provide a concise, structured tech news bulletin covering the most important developments from the last 24 hours."
    output_key: str = "newsFindings"
    tools: list = (google_search,)
    instruction: str = """
        Produce only a clean news bulletin.

        Task:
        Use google_search to collect the top 10 most important technology news from the last 24 hours.

        Coverage:
        - Programming (languages, frameworks, open-source updates)
        - DevTools (IDEs, AI coding, developer ecosystem)
        - Infrastructure (AWS, Azure, GCP, APIs, outages)
        - Big Tech (Google, Nvidia, Microsoft, Amazon, Meta, OpenAI, Apple)

        Rules:
        - Only include news within the last 24 hours
        - Use reputable sources and official blogs only
        - No commentary or reasoning in the output
        - Only bullet points with sources and links

        ### OUTPUT FORMAT
        Return ONLY a standard Python list of strings. Do not wrap in Markdown code blocks.
        
        Example Output Format:
        [
            "Storybook 10 ESM-only module automocking",
            "htmx 4.0 Alpha 1 release features",
            "Svelte November 2025 update features",
            "GitHub npm token security policy update"
        ]
    """

@dataclass
class DeepSearchConfig:
    agent_name: str = "DeepInvestigator"
    description: str = (
        "Perform a deep-dive investigation into the most recent tech news "
        "for a provided headline, strictly within the last 24 hours."
    )
    output_key: str = "deepNewsSearch"
    tools: list = (google_search,)
    instruction: str = """
        Role:
        You are an Expert Technical News Analyst specializing in timely,
        fact-verified technology reporting.

        Objective:
        You take will receive a paragraph list from {newsFindings} of trending Tech News Headline.
        Conduct in-depth research for every news topic and produce a concise and authoritative
        deep-dive summary for each topic (~350 words).

        Requirements:
        1. Confirm news recency — only include information published within
           the last 24 hours. Verify timestamps from reputable sources.
        2. Gather details from multiple credible sources (company press releases,
           major tech publishers, regulatory filings).
        5. Cross-check major facts across at least two independent sources.
        6. Avoid speculation — label clearly if inference is necessary.

        Coverage Scope:
        - Programming (languages, frameworks, open-source updates)
        - Developer Tools & AI (IDEs, copilots, productivity tools)
        - Infrastructure (Cloud providers, APIs, outages, chips)
        - Big Tech & Market Shifts (Google, Nvidia, Microsoft, Amazon,
          Meta, OpenAI, Apple)

        Output Rules:
        - Maintain a professional, neutral tone.
        - Include a structured clean output with:
            * A headline title field
            * ~250-word deep-dive summary
            * 3–5 bullet-point key facts
        - No internal commentary, chain-of-thought, or explanation of process.

        Denial Rule:
        If the headline does not have any credible updates in the last
        24 hours, do not respond just leave that news.
    """

@dataclass
class RootAgentConfig:
    agent_name: str = "NewsCoordinator"
    description: str = "Oversees the end-to-end workflow for technology news collection and deep investigative reporting."
    output_key: str = "deepNewsSummary"
    tools: list = ()
    instruction: str = """
        You are the Chief Technology News Editor responsible for coordinating a complete research and analysis pipeline to produce an authoritative deep-dive technology briefing.

        WORKFLOW:

        1. Fetch Headlines:
           Invoke the `NewsFindingAgent` tool.
           Request: "Retrieve the top 10 most important technology news headlines published within the last 24 hours."
           The tool will return a clean list of verified headlines.

        2. Perform Deep Analysis:
           Invoke the `DeepInvestigator` tool with the headline list obtained in Step 1.
           Request: "Conduct a full deep-dive investigation for each headline and produce a structured analytical report of no fewer than 350 words per topic."
           
           The investigator must:
           – Verify time-sensitivity and source credibility
           – Cross-reference facts across multiple reputable publishers
           – Deliver a professional, neutral, evidence-based report for every valid headline

        3. Final Delivery:
           Return the complete set of deep-dive reports produced by the `DeepInvestigator` as the final response to the user.
    """