pyproject.toml
[project]
name = "lucidtrend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.19.0",
    "python-dotenv>=1.2.1",
]
